Bugfix: don't cover content with menu on mobile
If the trigger is **inside** the container, the container will always be full
height as it contains the menu, which itself is full height.

This means it covers the body content.

So the container should be the element that is shown and hidden.

The trigger needs to be a sibling of the element that it shows and hides.
So we move the trigger outside the container.